Move SQL related annotations from data-jdbc to data-model
