+++ data-jdbc/src/test/groovy/io/micronaut/data/jdbc/oraclexe/jsonview/Student.java
@@ -4,7 +4,7 @@ import io.micronaut.data.annotation.GeneratedValue;
 import io.micronaut.data.annotation.Id;
 import io.micronaut.data.annotation.MappedEntity;
 import io.micronaut.data.annotation.Relation;
-import io.micronaut.data.jdbc.annotation.JoinTable;
+import io.micronaut.data.annotation.sql.JoinTable;
 
 import java.time.LocalDateTime;
 import java.util.Collections;
