+++ data-model/src/main/java/io/micronaut/data/annotation/sql/JoinTable.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.data.jdbc.annotation;
+package io.micronaut.data.annotation.sql;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -23,18 +23,33 @@ import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 /**
- * The repeatable container of @{@link io.micronaut.data.jdbc.annotation.JoinColumn}.
+ * Subset of the JPA join table annotation allowing support for bidirectional and unidirectional one-to-many join fetches only.
  *
- * @author Denis Stepanov
- * @since 2.4.0
+ * <p>Unlike the JPA version this is simplification and relies on a thid</p>
+ *
+ * @author graemerocher
+ * @since 1.0.0
  */
 @Target({METHOD, FIELD})
 @Retention(RUNTIME)
-public @interface JoinColumns {
+public @interface JoinTable {
+    /**
+     * @return The name of the join table
+     */
+    String name() default "";
 
     /**
-     * @return The join columns that map the relationship.
+     * @return The join columns to use.
      */
-    JoinColumn[] value();
+    JoinColumn[] joinColumns() default {};
 
+    /**
+     * @return The inverse join columns to use.
+     */
+    JoinColumn[] inverseJoinColumns() default {};
+
+    /**
+     * @return The alias to use for the query
+     */
+    String alias() default "";
 }
